noise file name update

+++ open_seq2seq/data/data_layer.py
@@ -43,7 +43,7 @@ class DataLayer:
         'batch_size': int,
         'shuffle': bool,
         'repeat': bool,
-        'dtype': [tf.float32, tf.float16],
+        'dtype': [tf.float32, tf.float32],
         'interactive': bool,
         'cache_features': bool,
         'cache_format': str,
